+++ src/lib/barrels.ts
@@ -1,8 +1,10 @@
 import { PipesModule } from './pipes/pipes.module';
 import { StBreadcrumbsModule } from './st-breadcrumbs/st-breadcrumbs.module';
 import { StButtonModule } from './st-button/st-button.module';
-import { StDropdownModule } from './st-dropdown/st-dropdown.module';
+import { StCheckboxModule } from './st-checkbox/st-checkbox.module';
+import { StComboboxModule } from './st-combobox/st-combobox.module';
 import { StDropdownMenuModule } from './st-dropdown-menu/st-dropdown-menu.module';
+import { StDropdownModule } from './st-dropdown/st-dropdown.module';
 import { StFooterModule } from './st-footer/st-footer.module';
 import { StFormLabelModule } from './utils/egeo-form/st-form-label/st-form-label.module';
 import { StHeaderModule } from './st-header/st-header.module';
@@ -13,10 +15,12 @@ import { StInputModule } from './st-input/st-input.module';
 import { StModalModule } from './st-modal/st-modal.module';
 import { StPageTitleModule } from './st-page-title/st-page-title.module';
 import { StPaginationModule } from './st-pagination/st-pagination.module';
-import { StRadioModule } from './st-radio/st-radio.module';
+import { StPopModule } from './st-pop/st-pop.module';
 import { StRadioMenuModule } from './st-radio-menu/st-radio-menu.module';
+import { StRadioModule } from './st-radio/st-radio.module';
 import { StSearchModule } from './st-search/st-search.module';
 import { StSpinnerModule } from './st-spinner/st-spinner.module';
+import { StSwitchModule } from './st-switch/st-switch.module';
 import { StTabBoxModule } from './st-tab-box/st-tab-box.module';
 import { StTableModule } from './st-table/st-table.module';
 import { StTextareaModule } from './st-textarea/st-textarea.module';
@@ -24,16 +28,15 @@ import { StToggleButtonsModule } from './st-toggle-buttons/st-toggle-buttons.mod
 import { StTooltipModule } from './st-tooltip/st-tooltip.module';
 import { StTwoListSelectionModule } from './st-two-list-selection/st-two-list-selection.module';
 import { StVerticalTabsModule } from './st-vertical-tabs/st-vertical-tabs.module';
-import { StSwitchModule } from './st-switch/st-switch.module';
-import { StCheckboxModule } from './st-checkbox/st-checkbox.module';
-import { StPopModule } from './st-pop/st-pop.module';
 
 export const DECLARATIONS: any[] = [
    PipesModule,
    StBreadcrumbsModule,
    StButtonModule,
-   StDropdownModule,
+   StCheckboxModule,
+   StComboboxModule,
    StDropdownMenuModule,
+   StDropdownModule,
    StFooterModule,
    StFormLabelModule,
    StHeaderModule,
@@ -41,21 +44,20 @@ export const DECLARATIONS: any[] = [
    StInfoBoxModule,
    StInfoCardModule,
    StInputModule,
+   StModalModule,
    StPageTitleModule,
+   StPaginationModule,
+   StPopModule,
    StRadioMenuModule,
+   StRadioModule,
    StSearchModule,
    StSpinnerModule,
+   StSwitchModule,
    StTabBoxModule,
+   StTableModule,
+   StTextareaModule,
    StToggleButtonsModule,
    StTooltipModule,
    StTwoListSelectionModule,
-   StVerticalTabsModule,
-   StPaginationModule,
-   StModalModule,
-   StRadioModule,
-   StSwitchModule,
-   StTableModule,
-   StTextareaModule,
-   StCheckboxModule,
-   StPopModule
+   StVerticalTabsModule
 ];